requirements_mine.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
defusedxml==0.5.0
Django==1.11.13
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-angular==2.0.3
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-compressor==2.2
django-filer==1.3.2
django-filter==1.1.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-mptt==0.8.7
django-parler==1.8
django-polymorphic==1.3
django-post-office==3.0.3
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-select2==6.0.2
-e git+https://github.com/rorschachrev/dj-pf@80b17fc8d032bb3c4d8696d261acccb4b3807aec#egg=django_shop&subdirectory=django-shop
django-treebeard==4.2.0
djangocms-admin-style==1.2.8
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-text-ckeditor==3.4.0
djangorestframework==3.7.7
djangoshop-paypal==0.2.1
djangoshop-stripe==0.12.3
drf-haystack==1.6.1
easy-thumbnails==2.5
elasticsearch==1.7.0
hiredis==0.2.0
html5lib==0.9999999
idna==2.6
jsonfield==1.0.3
libsass==0.12.3
oauthlib==1.0.3
olefile==0.45.1
paypalrestsdk==1.13.1
Pillow==4.3.0
pkg-resources==0.0.0
pycparser==2.18
pyOpenSSL==18.0.0
python-dateutil==2.6.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2015.7
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
rjsmin==1.0.12
six==1.11.0
stripe==1.53.0
Unidecode==0.4.18
urllib3==1.22
virtualenv==13.1.2
